Split the GPU hardware models into per-vendor files
hardware_type_utils.cc held both vendors' models, so one shared
GpuFlopCapabilities struct had to name its fields for both, and each function
carried its own vendor branch around a table lookup.

Move the NVIDIA tables to cuda_type_utils.cc and the AMD tables to
rocm_type_utils.cc, each owning a private capability struct. The NVIDIA one
goes back to cuda_core and tensor_core; the AMD one uses vector_unit and
matrix_unit. Both namespaces expose the same three PerCore signatures, so the
vendors are interchangeable at the call.

hardware_type_utils.cc keeps the public signatures and dispatches on
device_vendor, so callers are unchanged and the next vendor is a third file
plus one branch. ParseHardwareType stays there as it is vendor-neutral.

No behaviour change. The tables and the AMD gfx resolution moved verbatim, and
the existing tests pass unmodified since they exercise the public functions.
