+++ xprof/utils/hardware_type_utils.h
@@ -24,40 +24,7 @@ limitations under the License.
 namespace tensorflow {
 namespace profiler {
 
-struct GpuFlopCapabilities {
-  struct FlopCapabilityOnPrecisions {
-    double fp64_tflops = 0;
-    double fp32_tflops = 0;  // also for tf32 for nvidia tensor core
-    double bf16_tflops = 0;
-    double fp16_tflops = 0;
-    double fp8_tflops = 0;
-    double int8_tops = 0;
-    double fp4_tflops = 0;
-    double int4_tops = 0;
-
-    void ScaleWith(double scale) {
-      fp64_tflops *= scale;
-      fp32_tflops *= scale;
-      bf16_tflops *= scale;
-      fp16_tflops *= scale;
-      fp8_tflops *= scale;
-      int8_tops *= scale;
-      fp4_tflops *= scale;
-      int4_tops *= scale;
-    }
-  };
-
-  // Vector unit: NVIDIA = Cuda Cores, AMD = vector ALUs
-  FlopCapabilityOnPrecisions vector_unit;
-  // Matrix unit: NVIDIA = Tensor Cores, AMD = Matrix Cores (MFMA)
-  FlopCapabilityOnPrecisions matrix_unit;
-  bool has_matrix_unit_sparsity_support = false;
-
-  void ScaleWith(double scale) {
-    vector_unit.ScaleWith(scale);
-    matrix_unit.ScaleWith(scale);
-  }
-};
+// Dispatches to the vendor model in cuda_type_utils.h or rocm_type_utils.h.
 
 // Get peak single precision throughput of the GPU in GFLOPS per core (SM, CU)
 // TODO: Need design on how to use the sparsity capability of FLOPs.
